% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfinfo.R
\name{dfinfo}
\alias{dfinfo}
\title{Mutates a column info with the word "info}
\usage{
dfinfo(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a data frame
}
\description{
Mutates a column info with the word "info
}
