% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdownpractice-package.R
\docType{package}
\name{pkgdownpractice-package}
\alias{pkgdownpractice}
\alias{pkgdownpractice-package}
\title{pkgdownpractice: Playground For Github Workflows}
\description{
Practice playground for github workflows.
}
\author{
\strong{Maintainer}: Zelos Zhu \email{zelos.zhu@atorusresearch.com}

}
